refactor: user signup
