+++ src/api/users.js
@@ -1,8 +1,5 @@
 import { fetchJson } from '$lib/utils';
 
-export const signup = (name, email, password) =>
-	fetchJson('/api/users/signup', { name, email, password });
-
 const goHome = () => (location.href = '/');
 
 export const logout = async () => {
